+++ src/QuickMarcEditor/QuickMarcEditor.test.js
@@ -17,7 +17,7 @@ import { QUICK_MARC_ACTIONS } from './constants';
 import { MARC_TYPES } from '../common/constants';
 
 import Harness from '../../test/jest/helpers/harness';
-
+/* eslint-disable max-lines */
 jest.mock('react-router', () => ({
   ...jest.requireActual('react-router'),
   useLocation: () => ({